+++ NEWS.md
@@ -3,7 +3,10 @@
 # 0.1.9999 (dev)
 
 * reimplementation of image_collection_cube::read_chunk(), supporting image masks and custom gdalwarp arguments 
-
+* renamed filter_predicate -> filter_pixel
+* added gdalcubes namespace
+* removed reduce and stream commands from command line interface
+* added exec command to command line interface
 
 # 0.1.1